+++ src/br/edu/cafeteria/servico/Promocional.java
@@ -1,7 +1,5 @@
 package br.edu.cafeteria.servico;
 
-// import br.edu.cafeteria.modelo.Pedido;
-
 public interface Promocional {
-    public double aplicarDesconto(double porcentagem);
+    public float aplicarDesconto(float porcentagem);
 }
